refactor: eliminate duplicate download logic using useDownloadFile hook
Consolidate file download functionality by using the existing useDownloadFile hook instead of duplicating logic in components.

Changes:
- Enhanced useDownloadFile hook with better filename extraction and auth validation
- Refactored TrackedTask component to use useDownloadFile (removed ~50 lines)
- Refactored QueueCard component to use useDownloadFile (removed ~55 lines)
- Removed unnecessary imports (TokenStorage, apiClient, toast) from components

Benefits:
- Single source of truth for download logic
- Easier maintenance and consistency
- Improved code reusability
