+++ docs/context/README.md
@@ -1,0 +1,5 @@
+# Context
+
+This directory contains context documents that are useful during feature development or between pull requests. These are meant to be temporary documents used to hold context and notes for development and LLMS.
+
+These documents should be removed before being commited to the main brain (once properly using Git Flow).
